Move header-only c10/util container/storage utils to torch/headeronly
Migrate four already-header-only c10/util headers into torch/headeronly/util/,
leaving the old c10 paths as one-line forwarders (fully backward compatible, no
consumer include changes):

  Array.h, Bitset.h, FbcodeMaps.h, SmallBuffer.h

Part of the wave moving c10 headers that need no .cpp/libtorch linking into the
header-only library. Symbols are defined canonically in torch::headeronly and
re-aliased into c10 (the bit_cast.h convention); Bitset keeps its c10::utils
home as torch::headeronly::utils, and its hidden-friend operator== plus free
operator!= are re-exported so comparisons keep working via ADL.

c10/util/OptionalArrayRef.h was a container-chunk candidate but is deferred: it
wraps c10::ArrayRef, which is a subclass of HeaderOnlyArrayRef defined in
c10/util/ArrayRef.h (there is no torch::headeronly::ArrayRef), so it is not
header-only until ArrayRef itself is.

Test Plan:

Full build (confirms all libtorch consumers still compile):

```
pip install -e . --no-build-isolation
```

Header-only ABI check (compiles + runs without linking libtorch):

```
build/bin/test_aoti_abi_check --gtest_filter='TestArray*:TestBitset*:TestFbcodeMaps*:TestSmallBuffer*'
```

Lint:

```
lintrunner -a <changed files>
```

This PR was authored with the assistance of Claude (Claude Code).

[ghstack-poisoned]
